# docs/api.yaml
openapi: 3.0.0
info:
  title: 背了么 API
  description: |
    “背了么”是一款面向大学生的背单词App，提供标准背词、分级闯关、排行榜、AI词本等功能。
    
    ## 认证方式
    除登录注册外，所有接口需要在请求头中携带JWT Token：`Authorization: Bearer <token>`
    
    ## 统一响应格式
    - 成功: `{"code": 200, "message": "success", "data": {...}}`
    - 错误: `{"code": 400, "message": "错误信息", "errors": [...]}`
  version: 1.0.0
  contact:
    name: 背了么团队
  license:
    name: MIT

servers:
  - url: http://localhost:8080/api
    description: 开发环境
  - url: https://api.beileme.com/api
    description: 生产环境

tags:
  - name: Auth
    description: 用户认证相关接口
  - name: User
    description: 用户信息相关接口
  - name: Words
    description: 单词学习相关接口
  - name: Challenge
    description: 闯关对战相关接口
  - name: Leaderboard
    description: 排行榜相关接口
  - name: Wordbook
    description: AI词本相关接口
  - name: Study
    description: 学习统计相关接口

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
      required:
        - code
        - message

    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string

    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: zhangsan
        password:
          type: string
          minLength: 6
          example: "123456"
          writeOnly: true

    RegisterRequest:
      type: object
      required:
        - username
        - password
        - email
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: zhangsan
        password:
          type: string
          minLength: 6
          example: "123456"
          writeOnly: true
        email:
          type: string
          format: email
          example: zhangsan@example.com

    UserInfo:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          example: zhangsan
        avatar:
          type: string
          nullable: true
          example: https://cdn.beileme.com/avatars/1.jpg
        totalScore:
          type: integer
          example: 1250
        level:
          type: integer
          example: 3

    LoginResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          type: object
          properties:
            token:
              type: string
              example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
            userInfo:
              $ref: '#/components/schemas/UserInfo'

    Word:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        word:
          type: string
          example: abandon
        meaning:
          type: array
          items:
            type: string
          example: ["v. 放弃", "n. 放任"]
        phonetic:
          type: string
          example: "/əˈbændən/"
        example:
          type: array
          items:
            type: string
          example: ["He abandoned his plan."]

    DailyWordsResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          type: object
          properties:
            newWords:
              type: array
              items:
                $ref: '#/components/schemas/Word'
            reviewWords:
              type: array
              items:
                $ref: '#/components/schemas/Word'
            total:
              type: integer
              description: 当前返回的新词与复习词合计
              example: 12
            learnableWordCount:
              type: integer
              description: |
                可学单词数量（学会前都算待学习）：
                `word` 中尚未出现在该用户 `user_word` 的词数
                + `user_word` 中 status=0、today_mastered=0、review_count=0 的词数。
              example: 1800
            reviewableWordCount:
              type: integer
              description: |
                可复习单词数量：`user_word` 中 status=0、review_count>0 且 next_review<=NOW() 的词数（仅查询，不改库表）。
              example: 42

    LearnResponse:
      type: object
      properties:
        nextWord:
          $ref: '#/components/schemas/NextWordInfo'
        completedCount:
          type: integer
          description: 今日已完成单词数
          example: 3
        totalCount:
          type: integer
          description: 今日需要学习的单词总数
          example: 20

    NextWordInfo:
      type: object
      properties:
        wordId:
          type: integer
          format: int64
          example: 123
        word:
          type: string
          example: abandon
        meaning:
          type: array
          items:
            type: string
          example: ["v. 放弃", "n. 放任"]
        phonetic:
          type: string
          example: "/əˈbændən/"
        stage:
          type: integer
          minimum: 1
          maximum: 3
          description: "下一阶段：1-选义 2-选例 3-认词"
          example: 2
        options:
          type: array
          items:
            type: string
          description: "阶段1时的4个选项（干扰项）"
          example: ["放弃", "接受", "坚持", "离开"]
        example:
          type: string
          description: "阶段2时的例句"
          example: "He abandoned his plan." 

    LearnRequest:
      type: object
      required:
        - wordId
        - isCorrect
        - stage
      properties:
        wordId:
          type: integer
          format: int64
          example: 1
        isCorrect:
          type: boolean
          example: true
        stage:
          type: integer
          minimum: 1
          maximum: 3
          description: "学习阶段：1-选义 2-选例 3-认词"
          example: 1

    ChallengeStartRequest:
      type: object
      required:
        - levelType
      properties:
        levelType:
          type: integer
          minimum: 1
          maximum: 3
          description: "1-初级场 2-中级场 3-高级场"
          example: 1

    ChallengeQuestion:
      type: object
      properties:
        id:
          type: integer
          format: int64
        word:
          type: string
        options:
          type: array
          items:
            type: string
        correctIndex:
          type: integer

    ChallengeStartResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          type: object
          properties:
            challengeId:
              type: string
            questions:
              type: array
              items:
                $ref: '#/components/schemas/ChallengeQuestion'
            timeLimit:
              type: integer

    ChallengeSubmitRequest:
      type: object
      required:
        - challengeId
        - levelType
        - answers
      properties:
        challengeId:
          type: string
        levelType:
          type: integer
          minimum: 1
          maximum: 3
          description: "关卡等级（1初级 2中级 3高级）"
        answers:
          type: array
          items:
            type: object
            properties:
              questionId:
                type: integer
                format: int64
              selectedIndex:
                type: integer
              timeSpent:
                type: integer

    ChallengeSubmitResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          type: object
          properties:
            score:
              type: integer
            correctCount:
              type: integer
            totalCount:
              type: integer
            accuracy:
              type: number
            addedScore:
              type: integer
            totalScore:
              type: integer

    RankItem:
      type: object
      properties:
        rank:
          type: integer
        userId:
          type: integer
          format: int64
        username:
          type: string
        avatar:
          type: string
        totalScore:
          type: integer

    StudyStats:
      type: object
      properties:
        todayStudy:
          type: integer
        todayReview:
          type: integer
        totalWords:
          type: integer
        totalVocabulary:
          type: integer
          description: 当前词库总单词数（word表总量）
        masteredWords:
          type: integer
        wordbookWords:
          type: integer
        dueReviewCount:
          type: integer
        totalScore:
          type: integer
        level:
          type: integer

    StudyTrend:
      type: object
      properties:
        date:
          type: string
          format: date
        studyCount:
          type: integer
        reviewCount:
          type: integer
        correctRate:
          type: number

    StudyCalendar:
      type: object
      properties:
        year:
          type: integer
        month:
          type: integer
        studyDates:
          type: array
          description: 该月存在 study_record 的日期（yyyy-MM-dd）
          items:
            type: string
            format: date

    MemoryHintBlock:
      type: object
      description: 单个联想记忆模块；如果该单词不适合该类型，可返回 null
      properties:
        title:
          type: string
          description: 前端展示标题
          example: 谐音记忆
        content:
          type: string
          description: 联想核心内容
          example: 俺必胜
        explanation:
          type: string
          description: 联想与释义之间的关联说明
          example: 有雄心的人才会说“俺必胜”，帮助记住 ambition 表示雄心、野心。
      required:
        - title
        - content
        - explanation

    AIMemoryContent:
      type: object
      properties:
        word:
          type: string
          example: ambition
        meaning:
          type: string
          description: 当前联想主要绑定的中文释义
          example: 野心；雄心
        homophonic:
          allOf:
            - $ref: '#/components/schemas/MemoryHintBlock'
          nullable: true
          example:
            title: 谐音记忆
            content: 俺必胜
            explanation: 有雄心的人才会说“俺必胜”，帮助记住 ambition 表示雄心、野心。
        morpheme:
          allOf:
            - $ref: '#/components/schemas/MemoryHintBlock'
          nullable: true
          example:
            title: 拆词联想
            content: amb + it + ion
            explanation: 可以联想成四处走动拉选票，体现强烈进取心和野心。
        story:
          allOf:
            - $ref: '#/components/schemas/MemoryHintBlock'
          nullable: true
          example:
            title: 故事记忆
            content: 他为了竞选成功四处奔走，嘴里一直念着“俺必胜”，这份 ambition 让他停不下来。
            explanation: 通过竞选奔走和“俺必胜”的故事，把 ambition 和记成雄心、野心。
        summary:
          type: string
          description: 一句话记忆总结
          example: ambition = 俺必胜 = 有雄心、有野心的人
        notes:
          type: string
          nullable: true
          description: 生成说明；用于提示某些记忆类型为何缺失
          example: 该词谐音联想较自然，拆词为辅助记忆，并非严格词源分析。
      required:
        - word
        - meaning
        - summary

    AIMemoryResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/AIMemoryContent'
      required:
        - code
        - message
        - data

paths:
  /auth/register:
    post:
      tags:
        - Auth
      summary: 用户注册
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: 参数错误或用户名已存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/login:
    post:
      tags:
        - Auth
      summary: 用户登录
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: 用户名或密码错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/logout:
    post:
      tags:
        - Auth
      summary: 用户登出
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 登出成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /user/profile:
    get:
      tags:
        - User
      summary: 获取用户信息
      operationId: getUserProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/UserInfo'
    put:
      tags:
        - User
      summary: 更新用户信息
      operationId: updateUserProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                avatar:
                  type: string
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /user/password:
    put:
      tags:
        - User
      summary: 修改密码
      operationId: changePassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - oldPassword
                - newPassword
              properties:
                oldPassword:
                  type: string
                newPassword:
                  type: string
                  minLength: 6
      responses:
        '200':
          description: 修改成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /words/daily:
    get:
      tags:
        - Words
      summary: 获取今日单词
      operationId: getDailyWords
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyWordsResponse'

  /words/learn:
    post:
      tags:
        - Words
      summary: 提交学习结果
      description: |
        渐进式学习流程：
        - 阶段1（选义）：显示单词和4个中文选项，选对进入阶段2，选错回到阶段1
        - 阶段2（选例）：显示单词+例句，选择是否认识，选认识进入阶段3，选不认识回到阶段1
        - 阶段3（认词）：只显示单词，选择是否认识，选认识标记为今日掌握，选不认识回到阶段1
        提交后返回下一个需要学习的单词
      operationId: submitLearnResult
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LearnRequest'
      responses:
        '200':
          description: 提交成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearnResponse'

  /words/search:
    get:
      tags:
        - Words
      summary: 搜索单词
      operationId: searchWords
      security:
        - bearerAuth: []
      parameters:
        - name: keyword
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 搜索成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /words/{id}:
    get:
      tags:
        - Words
      summary: 获取单词详情
      operationId: getWordDetail
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Word'
        '404':
          description: 单词不存在

  /challenge/start:
    post:
      tags:
        - Challenge
      summary: 开始闯关
      operationId: startChallenge
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChallengeStartRequest'
      responses:
        '200':
          description: 开始成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeStartResponse'

  /challenge/submit:
    post:
      tags:
        - Challenge
      summary: 提交闯关结果
      operationId: submitChallenge
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChallengeSubmitRequest'
      responses:
        '200':
          description: 提交成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeSubmitResponse'

  /challenge/records:
    get:
      tags:
        - Challenge
      summary: 获取闯关记录
      operationId: getChallengeRecords
      security:
        - bearerAuth: []
      parameters:
        - name: levelType
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 3
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /leaderboard:
    get:
      tags:
        - Leaderboard
      summary: 获取排行榜
      operationId: getLeaderboard
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [total, daily, weekly]
            default: total
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      list:
                        type: array
                        items:
                          $ref: '#/components/schemas/RankItem'
                      myRank:
                        type: integer

  /wordbook/list:
    get:
      tags:
        - Wordbook
      summary: 获取生词本列表
      operationId: getWordbookList
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /wordbook/add/{wordId}:
    post:
      tags:
        - Wordbook
      summary: 添加单词到生词本
      operationId: addToWordbook
      security:
        - bearerAuth: []
      parameters:
        - name: wordId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 添加成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /wordbook/remove/{wordId}:
    delete:
      tags:
        - Wordbook
      summary: 从生词本移除
      operationId: removeFromWordbook
      security:
        - bearerAuth: []
      parameters:
        - name: wordId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 移除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /wordbook/ai/{wordId}:
    get:
      tags:
        - Wordbook
      summary: 获取AI联想记忆内容
      description: |
        为指定单词生成结构化的联想记忆内容，优先返回更适合该单词的记忆钩子：
        - 谐音记忆：将英文发音与中文短语建立记忆锚点
        - 拆词联想：基于词形拆分、词根词缀或近似结构进行辅助记忆
        - 故事记忆：用简短场景把单词含义串起来

        说明：
        - 并非每个单词都适合所有类型；不适合时对应字段可为 null
        - 拆词联想用于辅助记忆，不保证严格词源学分析
      operationId: getAIMemoryContent
      security:
        - bearerAuth: []
      parameters:
        - name: wordId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIMemoryResponse'
              example:
                code: 200
                message: success
                data:
                  word: ambition
                  meaning: 野心；雄心
                  homophonic:
                    title: 谐音记忆
                    content: 俺必胜
                    explanation: 有雄心的人才会说“俺必胜”，帮助记住 ambition 表示雄心、野心。
                  morpheme:
                    title: 拆词联想
                    content: amb + it + ion
                    explanation: 可以联想成四处走动拉选票，体现强烈进取心和野心。
                  story:
                    title: 故事记忆
                    content: 他为了竞选成功四处奔走，嘴里一直念着“俺必胜”，这份 ambition 让他停不下来。
                    explanation: 通过竞选奔走和“俺必胜”的故事，把 ambition 和记成雄心、野心。
                  summary: ambition = 俺必胜 = 有雄心、有野心的人
                  notes: 该词谐音联想较自然，拆词为辅助记忆，并非严格词源分析。
        '404':
          description: 单词不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: AI生成失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /study/stats:
    get:
      tags:
        - Study
      summary: 获取学习统计
      operationId: getStudyStats
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/StudyStats'

  /study/trend:
    get:
      tags:
        - Study
      summary: 获取学习趋势
      operationId: getStudyTrend
      security:
        - bearerAuth: []
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 7
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/StudyTrend'

  /study/calendar:
    get:
      tags:
        - Study
      summary: 获取学习日历
      operationId: getStudyCalendar
      security:
        - bearerAuth: []
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
        - name: month
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 12
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/StudyCalendar'